requirements.txt
streamlit==1.32.0
google-cloud-aiplatform==1.42.1
face-recognition==1.3.0
mediapipe==0.10.14
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4 